refactor/const
