+++ src/components/api.js
@@ -67,63 +67,3 @@ export default class Api {
     return this._setRequest('/users/me/avatar', 'PATCH', { avatar: imageUrl })
   }
 }
-
-// Шаблон обработчика запроса на сервер
-const requestConfig = async (url, method, data, headers = apiConfig.headers) => {
-  const res = await fetch(`${apiConfig.serverUrl}${url}`, {
-    method: method,
-    headers: headers,
-    body: JSON.stringify(data)
-  })
-  return checkConnect(res)
-}
-
-// Запрос на сервер:
-const checkConnect = res => {
-  if (!res.ok) {
-    return Promise.reject(`Ошибка ${res.status}`)
-  }
-  return res.json()
-}
-
-// Получить карточки
-const getCards = () => {
-  return requestConfig('/cards')
-}
-
-// Получить данные пользователя
-const getUserId = () => {
-  return requestConfig('/users/me')
-}
-
-// Отправить данные пользователя на сервер
-const editProfileData = (user) => {
-  return requestConfig('/users/me', 'PATCH', user)
-}
-
-// Отправить данные новой карточки на сервер
-const addNewCard = (newCard) => {
-  return requestConfig('/cards', 'POST', newCard)
-}
-
-// Удалить карточку добавленную пользователем
-const delNewCard = (cardId) => {
-  return requestConfig(`/cards/${cardId}`, 'DELETE')
-}
-
-// Добавить лайк карточке
-const addLikeCard = (cardId) => {
-  return requestConfig(`/cards/likes/${cardId}`, 'PUT')
-}
-
-// Удалить лайк карточки
-const delLikeCard = (cardId) => {
-  return requestConfig(`/cards/likes/${cardId}`, 'DELETE')
-}
-
-// Обновить аватар пользователя
-const refreshAvatar = (imageUrl) => {
-  return requestConfig('/users/me/avatar', 'PATCH', { avatar: imageUrl })
-}
-
-export { getUserId, getCards, editProfileData, addNewCard, delNewCard, addLikeCard, delLikeCard, refreshAvatar };
